+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Mail:
 * Roundcube is updated to version 1.2.0.
 * SSLv3 and RC4 are now no longer supported in incoming and outgoing mail (SMTP port 25).
 
+System:
+
+* fail2ban jails added for SMTP submission, Roundcube, ownCloud, the control panel, and munin.
+
 v0.18c (June 2, 2016)
 ---------------------